requirements.txt
requests==2.32.3
python-dotenv==1.0.1
pytest==9.0.2
pytest-mock==3.10.0